supplemental change for pull request 1973
